Lab1/Task2/Wine_predictor/requirements.txt
hopsworks
joblib
scikit-learn==1.1.1
great_expectations
pandas
matplotlib
seaborn
pillow